began implementing better naming convention for controller files
Localhost:4000/videoswap?FILENAME will spinup a browser, download file and close the browser in 10 seconds.

LESSONS LEARNED:
When trying to simulate the download, most of my issues came from the fact that the download was happening outside of the scope of Seleniums event listeners (because it was essentially a hack). I have yet to discover how to have Selenium track when a file is downloaded so that I can use Selenium event listeners to leverage the synchronous nature of it’s commands. That is why a set timeout, with an arbitrary time associated with it make is seem as though the download and browser.close() is synchronous.

res.end() has not been implemented yet so as the request hangs, it will perform the automation twice.