Added res.end().
LESSON LEARNED:
It only has bearing on the node server (localhost:4000) and ends the request for that page. the selenium browser instance processes continues as normal  even after the node request has been closed.

POTENTIAL BUG.
For the first time, the browser closed before the download was complete, leaving a file that was essential unusable in the default download location(/Downloads). I adjust the setTimeout for browserClose to be 20sec