autoMate is able to automate video swap process up until injecting the video and thumbnail into the add video dialogue.
LESSONS LEARNED
-learned to deal with promises

 var getCurrentURL=browser.getCurrentUrl();//with return a promise
    getCurrentURL.then(function(url){//resolved value can be accessed in then callback
        var fileName=dom.getFileName(url);
        res.locals.thumbnail=fileName;
        help.downloadAsset(browser,url);
        uploadVideoToControlRoom(browser,res); //deal with the asynchronous nature of node by passing the next step as a callback to this callback. This ensures that by the time this step is finished all of the values are populated.
    });

GOALS: Learn to use promises more judiciously. They are powerful especially in node.

Node is async as fuck and this project is helping me to see that. When wait times ca be as much as 500 miliseconds or more, values not readily available when assigned to a variable become difficult to manage. Thats why selenium web drivers promised-centric library makes sense. Really interesting.

The application is automates